requirements.txt
watchdog>=4.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
APScheduler>=3.10.0
pytest>=8.0.0
